% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Adolescent alcohol use data}
\format{A data frame with 663 rows:
\describe{
  \item{heavyalcohol}{Binge drinking}
  \item{peer}{Number of friends who drank alcohol}
}}
\usage{
alcohol
}
\description{
A simulated dataset containing data on adolecsent alcohol use.
}
\keyword{datasets}
