% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intmed.R
\docType{package}
\name{intmed}
\alias{intmed}
\title{intmed: A package for estimating the interventional effect for mediation analysis.}
\description{
intmed: A package for estimating the interventional effect for mediation analysis.
}
