% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med5.R
\name{mediate}
\alias{mediate}
\title{Estimating the mediation analysis based on the interventional effect}
\usage{
mediate(
  y,
  med,
  treat,
  mod = NULL,
  c = NULL,
  ymodel,
  mmodel,
  treat_lv = 1,
  control_lv = 0,
  incint = NULL,
  inc_mmint = FALSE,
  data,
  sim = 1000,
  conf.level = 0.95,
  out_scale = "difference",
  complete_analysis = FALSE
)
}
\arguments{
\item{y}{The outcome variable.}

\item{med}{A vector of the mediators.}

\item{treat}{The exposure variable.}

\item{mod}{A vector of moderators for moderated mediation analysis - Not yet implemented.}

\item{c}{A vector of covariates.}

\item{ymodel}{A character string specifying the outcome model. Current options are "regression" (for continuous variable), "logistic regression" (for binary variable), and "poisson regression" (for count variable)}

\item{mmodel}{A vector of character string specifying the mediator models. Current options are "regression" (for continuous variable), "logistic regression" (for binary variable), and "poisson regression" (for count variable)}

\item{treat_lv}{Value of the treatment variable used as the treatment condition. Default is 1.}

\item{incint}{A vector of boolean specifying if the exposure-mediator interactions are included into the outcome model. Default is NULL.}

\item{inc_mmint}{A boolean value specifying if the mediator-mediator interactions are included. Default is FALSE.}

\item{data}{A data frame containing all the analysis variables.}

\item{sim}{A numerical value specifying the number of simulation. Default is 1000.}

\item{conf.level}{A numerical value specifying the confidence interval the the estimates. Default is 0.95}

\item{out_scale}{A string specifying the scale of the analysis. Only difference scale is implemented.}

\item{complete_analysis}{Multiple imputation will be used to fill in missing value. Setting this flag to FALSE will force the analysis to be conducted on complete data.}

\item{contron_lv}{Value of the treatment variable used as the control condition. Default is 0.}
}
\description{
Estimating the mediation analysis based on the interventional effect
}
