% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{A simulated dataset for demonstrating mediation analysis with a single mediator}
\format{
A data frame with 1000 rows:
\describe{
  \item{x}{Exposure variable}
  \item{m}{Mediator}
  \item{y}{Outcome}
}
}
\usage{
sim_data
}
\description{
A simulated dataset with three variables.
}
\keyword{datasets}
