% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance.R
\docType{data}
\name{substance}
\alias{substance}
\title{Ficticious family intervention data}
\format{A data frame with 553 rows:
\describe{
  \item{sub_misuse}{Substance misuse in young adulthood. 1: Yes; 0: No}
  \item{fam_int}{Paricipation in family intervention during adolescence. 1: Yes; 0: No}
}}
\usage{
substance
}
\description{
A ficticious dataset about family intervention and youth substance misuse
}
\keyword{datasets}
